Values to convert from ft to m ma also be decimal numbers

+++ src/map/IMG/lblfile.cpp
@@ -42,7 +42,7 @@ static QString capitalized(const QString &str)
 static QByteArray ft2m(const QByteArray &str)
 {
 	bool ok;
-	int number = str.toInt(&ok);
+	double number = str.toDouble(&ok);
 	return ok ? QByteArray::number(qRound(number * 0.3048)) : str;
 }
